% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_consolidation.R
\name{assess_mmr_status}
\alias{assess_mmr_status}
\title{Assess MMR status based on IHC biomarker scores}
\usage{
assess_mmr_status(biomarkers_data)
}
\arguments{
\item{biomarkers_data}{data frame with biomarker scores, which must include the following columns:
"mlh1", "msh2", "msh6", "pms2". The function will add a new column "mmr_ihc_4" with the MMR status based on the following rules:
\itemize{
\item If at least one of MSH6, PMS2, MLH1, MSH2 is absent, MMR status is 'deficient'.
\item If at least one of MSH6, PMS2 is unknown (i.e., "Unk"), MMR status is 'Unk'.
\item If all of MSH6, PMS2 are present and none of MSH6, PMS2, MLH1, MSH2 is absent,
MMR status is 'intact'.
}}
}
\value{
A data frame with an additional column "mmr_ihc_4" containing the MMR status ("deficient", "intact", or "Unk") for each row.
}
\description{
Assess MMR status based on IHC biomarker scores
}
\examples{
library(TMAtools)
# create example biomarker data
example_data <- data.frame(
  mlh1 = c("present", "absent", "present"),
  msh2 = c("present", "present", "present"),
  msh6 = c("present", "present", "Unk"),
  pms2 = c("present", "present", "present"),
  stringsAsFactors = FALSE
)
assess_mmr_status(example_data)
}
