% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{consolidate_scores}
\alias{consolidate_scores}
\title{Consolidate biomarkers}
\usage{
consolidate_scores(
  biomarkers_file = NULL,
  biomarker_rules_file = NULL,
  output_file = NULL,
  biomarkers_data = NULL
)
}
\arguments{
\item{biomarkers_file}{Path to the Excel file containing biomarker data.}

\item{biomarker_rules_file}{Path to spreadsheet containing the consolidation rules for all biomarkers.
It must contain a sheet named "consolidation" with columns
"biomarker", "rule_type", "rule_value", "consolidated_value".}

\item{output_file}{Optional path to the output file. If NULL, the function will not save the output.}

\item{biomarkers_data}{Optinally, pass a data.frame or tibble with biomarker data
instead of passing \code{biomarkers_file}. Used for reconsolidation in \code{tmatools()}}
}
\value{
A data frame with translated biomarker scores.
}
\description{
This function consolidates biomarker scores for each case.
}
\examples{
library(TMAtools)
# grab folder with example TMA datasets
tma_dir <- system.file("extdata", "tma1", package = "TMAtools")
# define output files
combined_tma_file <- "combined_tma.xlsx"
deconvoluted_tma_file <- "deconvoluted_tma.xlsx"
translated_tma_file <- "translated_tma.xlsx"
consolidated_tma_file <- "consolidated_tma.xlsx"

# combine TMA datasets
combine_tma_spreadsheets(
 tma_dir = tma_dir,
 output_file = combined_tma_file
)

# deconvolute combined TMA dataset
deconvolute(
   tma_file = combined_tma_file,
   output_file = deconvoluted_tma_file
)

# translate numerical scores to nominal scores
translate_scores(
   biomarkers_file = deconvoluted_tma_file,
   biomarker_rules_file = "tmp/biomarker_rules.xlsx",
   output_file = translated_tma_file
)

# and consolidate nominal scores for each case
consolidated_data <- translate_and_consolidate_scores(
  biomarkers_file = deconvoluted_tma_file,
  required_biomarkers = c("ER", "TP53")
)
head(consolidated_data)
}
