% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{consolidate_single_patient}
\alias{consolidate_single_patient}
\title{Assess consolidated score per patient}
\usage{
consolidate_single_patient(
  rules_df,
  scores,
  unknown_values = c("Unk", "x"),
  late_na_ok = FALSE
)
}
\arguments{
\item{rules_df}{data.frame with consolidation rules for a single biomarker}

\item{scores}{character vector with all scores for a given biomarker/patient}

\item{late_na_ok}{If TRUE, NAs do not trigger error. Used for re-consolidation.}

\item{unknow_values}{character vector with values treated as unknown. Defaults to \code{c("Unk", "x")}.}
}
\value{
consolidated value (as character).
}
\description{
Assess consolidated score per patient
}
\keyword{internal}
