% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{deconvolute}
\alias{deconvolute}
\title{Deconvolute a combined TMA spreadsheet}
\usage{
deconvolute(tma_file, metadata = NULL, output_file = NULL, tma_id = NULL)
}
\arguments{
\item{tma_file}{Path to input Excel spreadsheet.
Both .xlsx and .xls formats are supported.
Must contain one "TMA map" sheet and one or more
biomarker-specific sheets (eg, output file of \code{combine_tma_spreadsheets()}).}

\item{metadata}{Optional \code{data.frame} with at least \code{core_id} and \code{accession_id} columns,
plus any other metadata columns that should be added to the output
(eg, patient age, sex, etc.).}

\item{output_file}{Optional path to output
Excel spreadsheet, which can be used as input to \code{translate_scores()}.}

\item{tma_id}{Optional TMA identifier (column \code{tma_id} will be added to output).}
}
\value{
A data frame with the deconvoluted data.
}
\description{
Reads TMA spreadsheet from \code{tma_file}
which must contain one "TMA map" sheet and one or more
biomarker-specific sheets.
}
\details{
The function will
match core IDs from the TMA map sheet
with the biomarker-specific sheets and return a
data frame with a "core_id" column and as many
columns for each biomarker as necessary.

For instance, if core ID 1 has 3 values for biomarker A,
the output will contain 3 columns for biomarker A
(A.c1, A.c2, A.c3); if another core ID has 2 values for
biomarker A, its corresponding A.c1 and A.c2 columns will
be filled with the values for that core ID and the
A.c3 column will be filled with \code{NA}.

If there are no values for a given core ID in a
biomarker-specific sheet, the corresponding
columns will all be filled with \code{NA}.
}
\examples{
library(TMAtools)
# grab folder with example TMA datasets
tma_dir <- system.file("extdata", "tma1", package = "TMAtools")
# define output files
combined_tma_file <- "combined_tma.xlsx"
deconvoluted_tma_file <- "deconvoluted_tma.xlsx"

# combine TMA datasets
combine_tma_spreadsheets(
 tma_dir = tma_dir,
 output_file = combined_tma_file,
 biomarker_sheet_index = 2,
 valid_biomarkers = c("ER", "p53") # optional, but recommended to avoid misspelling errors
)

# deconvolute combined TMA dataset
deconvoluted_data <- deconvolute(
   tma_file = combined_tma_file,
   output_file = deconvoluted_tma_file
)
print(deconvoluted_data)
}
