% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template_biomarker_rules.R
\name{get_biomarker_rules_template}
\alias{get_biomarker_rules_template}
\title{Get template biomarker rules file}
\usage{
get_biomarker_rules_template(output_file = "biomarker_rules_template.xlsx")
}
\arguments{
\item{output_file}{Path to output file. The template will be copied to this path.}
}
\description{
Get template biomarker rules file
}
