% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate.R
\name{get_consolidation_rules_df}
\alias{get_consolidation_rules_df}
\title{Get consolidation data.frame from biomarker rules file}
\usage{
get_consolidation_rules_df(biomarker_rules_file)
}
\arguments{
\item{biomarker_rules_file}{Path to spreadsheet containing the consolidation rules for all biomarkers.
It must contain a sheet named "consolidation" with columns
"biomarker", "rule_type", "rule_value", "consolidated_value".}
}
\value{
a data.frame of consolidation rules for each biomarker.
}
\description{
Get consolidation data.frame from biomarker rules file
}
