% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{get_default_biomarkers}
\alias{get_default_biomarkers}
\title{Get default biomarkers}
\usage{
get_default_biomarkers()
}
\value{
A character vector of default biomarker names.
}
\description{
This function returns a vector of default biomarkers commonly used in TMA analysis.
}
\examples{
library(TMAtools)
default_biomarkers <- get_default_biomarkers()
print(default_biomarkers)
}
