% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_and_consolidate.R
\name{get_default_dict}
\alias{get_default_dict}
\title{Get default biomarker dictionary}
\usage{
get_default_dict(biomarker)
}
\arguments{
\item{biomarker}{The name of the biomarker.}
}
\value{
A named vector representing the dictionary for the specified biomarker.
}
\description{
This function returns a dictionary for translating numerical scores to nominal scores for default biomarkers.
}
\examples{
options(width = 10000) # for printing
library(TMAtools)
biomarkers <- c("ER", "TP53", "CTNNB1", "WT1")
for (biomarker in biomarkers) {
   print(c(biomarker, get_default_dict(biomarker)))
}
}
