% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{get_dict_default}
\alias{get_dict_default}
\title{Get biomarker dictionary (default function)}
\usage{
get_dict_default(biomarker)
}
\arguments{
\item{biomarker}{The name of the biomarker.}
}
\value{
A named character vector representing the dictionary for the specified biomarker.
}
\description{
This function returns a dictionary (named character vector) for
translating numerical scores to nominal scores for default biomarkers.
}
\examples{
library(TMAtools)
biomarkers <- c("ER", "TP53", "CTNNB1", "WT1")
for (biomarker in biomarkers) {
   print(
      c(biomarker, get_dict_default(biomarker)), 
      width = 10000  # for prettier printing
   )
}
}
