% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_and_consolidate.R
\name{get_leopard_dict}
\alias{get_leopard_dict}
\title{Get dictionary for LEOPARD biomarkers (2025)}
\usage{
get_leopard_dict(biomarker)
}
\arguments{
\item{biomarker}{The name of the biomarker.}
}
\value{
A named vector representing the dictionary for the specified biomarker.
}
\description{
This function returns a dictionary for translating numerical scores to nominal scores for LEOPARD biomarkers.
}
\examples{
library(TMAtools)
print(get_leopard_dict("ER"))
print(get_leopard_dict("CTNNB1"))
}
