% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{get_translation_dictionary}
\alias{get_translation_dictionary}
\title{Get translation dictionary from biomarker rules file}
\usage{
get_translation_dictionary(biomarker_rules_file)
}
\arguments{
\item{biomarker_rules_file}{Path to spreadsheet containing the translation rules for all biomarkers.
It must contain a sheet named "translation" with columns
"biomarker", "original_score", "translated_score".}
}
\value{
data.frame with the translation rules.
}
\description{
Get translation dictionary from biomarker rules file
}
\keyword{internal}
