% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_spreadsheets.R
\name{read_data}
\alias{read_data}
\title{Read data from excel spreadsheet}
\usage{
read_data(input_path)
}
\arguments{
\item{input_path}{Path to input Excel spreadsheet}
}
\value{
a data.frame
}
\description{
Read data from excel spreadsheet
}
