% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmatools.R
\name{tmatools}
\alias{tmatools}
\title{TMAtools pipeline}
\usage{
tmatools(
  tma_dirs,
  biomarker_rules_file,
  output_dir = "tmatools_output",
  combined_tma_file = "1_combined_tma.xlsx",
  deconvoluted_tma_file = "2_deconvoluted_tma.xlsx",
  translated_tma_file = "3_translated_tma.xlsx",
  consolidated_tma_file = "4_consolidated_tma.xlsx",
  final_tma_file = "5_final_consolidated_tmas.xlsx",
  biomarker_sheet_index = 2,
  required_biomarkers = c("ER", "TP53")
)
}
\arguments{
\item{tma_dirs}{A character vector of TMA directory paths.}

\item{output_dir}{The directory where the output files will be saved.}

\item{combined_tma_file}{The name of the combined TMA file.}

\item{deconvoluted_tma_file}{The name of the deconvoluted TMA file.}

\item{consolidated_tma_file}{The name of the consolidated TMA file.}

\item{final_tma_file}{The name of the final file containing the consolidated scores from multiple TMAs processed.}

\item{biomarker_sheet_index}{The index of the biomarker sheet in the TMA file.}

\item{required_biomarkers}{A character vector of required biomarkers.}
}
\description{
This function runs the TMAtools pipeline for multiple TMA directories.
It combines TMA datasets, deconvolutes the combined dataset,
and translates numerical biomarker scores to nominal scores.
}
\examples{
library(TMAtools)
tma_dirs <- c(
 system.file("extdata", "tma1", package = "TMAtools"),
 system.file("extdata", "tma2", package = "TMAtools")
)
# Run the TMAtools pipeline
tmatools(
  tma_dirs = tma_dirs,
  output_dir = "tmatools_output"
)
}
