% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_and_consolidate.R
\name{translate_and_consolidate_scores}
\alias{translate_and_consolidate_scores}
\title{Consolidate biomarker scores}
\usage{
translate_and_consolidate_scores(
  biomarkers_file,
  required_biomarkers = NULL,
  get_dict = NULL,
  output_file = NULL
)
}
\arguments{
\item{biomarkers_file}{Path to the Excel file containing biomarker data.}

\item{required_biomarkers}{A vector of required biomarkers. If NULL, default biomarkers are used.}

\item{get_dict}{A function that returns a dictionary for translating numerical scores to nominal scores. If NULL, a default function is used.}

\item{output_file}{Optional path to the output file. If NULL, the function will not save the output.}
}
\value{
A data frame with consolidated biomarker scores and MMR status.
}
\description{
This function translates numerical scores of biomarkers to nominal scores and consolidates them for each case.
}
\examples{
library(TMAtools)
# grab folder with example TMA datasets
tma_dir <- system.file("extdata", "tma1", package = "TMAtools")
# define output files
combined_tma_file <- "combined_tma.xlsx"
deconvoluted_tma_file <- "deconvoluted_tma.xlsx"
consolidated_tma_file <- "consolidated_tma.xlsx"

# combine TMA datasets
combine_tma_spreadsheets(
 tma_dir = tma_dir,
 output_file = combined_tma_file
)

# deconvolute combined TMA dataset
deconvolute(
   tma_file = combined_tma_file,
   output_file = deconvoluted_tma_file
)

# translate numerical biomarker scores to nominal scores
# and consolidate them for each case
consolidated_data <- translate_and_consolidate_scores(
  biomarkers_file = deconvoluted_tma_file,
  required_biomarkers = c("ER", "TP53")
)
head(consolidated_data)
}
\seealso{
\code{\link{get_default_dict}} for the default dictionary function.
}
