% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_scores}
\alias{translate_scores}
\title{Translate biomarker scores from numerical to nominal}
\usage{
translate_scores(
  biomarkers_file,
  biomarker_rules_file = NULL,
  output_file = NULL
)
}
\arguments{
\item{biomarkers_file}{Path to the Excel file containing biomarker data.}

\item{biomarker_rules_file}{Path to spreadsheet containing the translation rules for all biomarkers.
It must contain a sheet named "translation" with columns
"biomarker", "original_score", "translated_score".}

\item{output_file}{Optional path to the output file. If NULL, the function will not save the output.}
}
\value{
A data frame with translated biomarker scores.
}
\description{
This function translates numerical scores of biomarkers to nominal scores.
}
\examples{
library(TMAtools)
# grab folder with example TMA datasets
tma_dir <- system.file("extdata", "tma1", package = "TMAtools")
# define output files
combined_tma_file <- "combined_tma.xlsx"
deconvoluted_tma_file <- "deconvoluted_tma.xlsx"
consolidated_tma_file <- "consolidated_tma.xlsx"

# combine TMA datasets
combine_tma_spreadsheets(
 tma_dir = tma_dir,
 output_file = combined_tma_file
)

# deconvolute combined TMA dataset
deconvolute(
   tma_file = combined_tma_file,
   output_file = deconvoluted_tma_file
)

# translate numerical biomarker scores to nominal scores
# and consolidate them for each case
consolidated_data <- translate_and_consolidate_scores(
  biomarkers_file = deconvoluted_tma_file,
  required_biomarkers = c("ER", "TP53")
)
head(consolidated_data)
}
