% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tma_spreadsheets.R
\name{combine_tma_spreadsheets}
\alias{combine_tma_spreadsheets}
\title{Combine TMA spreadsheets}
\usage{
combine_tma_spreadsheets(
  tma_dir,
  output_file = "combined_tma_spreadsheet.xlsx",
  biomarker_sheet_index = 2,
  tma_map_sheet_index = 1
)
}
\arguments{
\item{tma_dir}{The directory containing the TMA spreadsheets.}

\item{output_file}{The name of the output file.}

\item{biomarker_sheet_index}{The index of the sheet used for ALL files for the biomarker scores.}

\item{tma_map_sheet_index}{The index of the sheet used for ALL files for the TMA map.}
}
\value{
List of data frames (invisible).
}
\description{
This function combines multiple TMA spreadsheets into a single spreadsheet
that can be used for deconvolution with \code{TMAtools::deconvolute()}.
}
\examples{
library(TMAtools)
tma_dir <- system.file("extdata", package = "TMAtools")
combine_tma_spreadsheets(
    tma_dir = tma_dir,
    output_file = "combined_tma_spreadsheet.xlsx",
    biomarker_sheet_index = 2,
    tma_map_sheet_index = 1
)
# check the output Excel file, which should then be used as
# the input for `TMAtools::deconvolute()`
}
