% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute.R
\name{deconvolute}
\alias{deconvolute}
\title{Deconvolute TMA map}
\usage{
deconvolute(input_path)
}
\arguments{
\item{input_path}{Path to input Excel spreadsheet.
Both .xlsx and .xls formats are supported.
Must contain one "TMA map" sheet and one or more
biomarker-specific sheets.}
}
\value{
A data frame with the deconvoluted data.
The function will
match core IDs from the TMA map sheet
with the biomarker-specific sheets and return a
data frame with a "core_id" column and as many
columns for each biomarker as necessary.

For instance, if core ID 1 has 3 values for biomarker A,
the output will contain 3 columns for biomarker A
(A.c1, A.c2, A.c3); if another code ID has 2 values for
biomaker A, its corresponding A.c1 and A.c2 columns will
be filled with the values for that core ID and the
A.c3 column will be filled with NA.

If there are no values for a given core ID in a
biomarker-specific sheet, the corresponding
columns will all be filled with NA.
}
\description{
Reads TMA spreadsheet from \code{input_path}
which must contain one "TMA map" sheet and one or more
biomaker-specific sheets.
}
\examples{
library(TMAtools)
input_path <- system.file("extdata", "example.xlsx", package = "TMAtools")
deconvoluted_data <- deconvolute(input_path)
head(deconvoluted_data)
}
