% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmatools.R
\name{tmatools}
\alias{tmatools}
\title{TMAtools pipeline}
\usage{
tmatools(
  tma_dirs,
  biomarker_rules_file,
  output_dir = "tmatools_output",
  combined_tma_file = "1_combined_tma.xlsx",
  deconvoluted_tma_file = "2_deconvoluted_tma.xlsx",
  translated_tma_file = "3_translated_tma.xlsx",
  consolidated_tma_file = "4_consolidated_tma.xlsx",
  final_tma_file = "5_final_consolidated_tmas.xlsx",
  biomarker_sheet_index = 2
)
}
\arguments{
\item{tma_dirs}{A character vector of TMA directory paths.
Each directory must contain:
\itemize{
\item \verb{Score sheets}: one or more Excel files with biomarker scores (one biomarker per sheet).
\item \verb{Clean map}: an Excel file with "clean_map" in the file name.
This file corresponds to the sector map of your TMA that only contains
the core IDs within the corresponding cells. No other annotation outside
the map is allowed, as \code{TMAtools} uses the exact positions of core IDs
to map corresponding scores.
Optional:
\item \code{Metadata}: An Excel file with "metadata" in the name.
In a single tab, it must contain at least two columns:
"core_id" (core IDs that appear in the sector map) and "accession_id"
(case or patient identifiers). Optionally, you can add other columns
with additional metadata (e.g., age, sex, histotype, block number),
which will be carried forward to your output files.
}}

\item{biomarker_rules_file}{Path to spreadsheet containing the consolidation rules for all biomarkers.
It must contain a sheet named "consolidation" with columns
"biomarker", "rule_type", "rule_value", "consolidated_value".
It must not be located within any of the TMA directories being processed.}

\item{output_dir}{The directory where the output files will be saved.}

\item{combined_tma_file}{The name of the combined TMA file.}

\item{deconvoluted_tma_file}{The name of the deconvoluted TMA file.}

\item{translated_tma_file}{The name of the translated TMA file.}

\item{consolidated_tma_file}{The name of the consolidated TMA file.}

\item{final_tma_file}{The name of the final file containing the consolidated scores from multiple TMAs processed.}

\item{biomarker_sheet_index}{The index of the biomarker sheet in the TMA file.
All TMA files must have the biomarker data in the same sheet index. Defaults to 2 (ie, second sheet).}
}
\description{
This function runs the TMAtools pipeline for multiple TMA directories.
It combines TMA datasets, deconvolutes the combined dataset,
and translates numerical biomarker scores to nominal scores.
}
\examples{
library(TMAtools)
tma_dirs <- c(
 system.file("extdata", "tma1", package = "TMAtools"),
 system.file("extdata", "tma2", package = "TMAtools")
)
# spreadsheet with translation and consolidation rules
biomarker_rules_file <- system.file(
  "extdata", "biomarker_rules_example.xlsx",
  package = "TMAtools"
)
# Run the TMAtools pipeline
tmatools(
  tma_dirs = tma_dirs,
  biomarker_rules_file = biomarker_rules_file,
  output_dir = "tmatools_output"
)
}
